split structs and methods into their own files
