+++ Event.go
@@ -1,0 +1,7 @@
+package main
+
+type Event struct {
+	Description string
+	Outcome string
+	Triggered bool
+}
