+++ Interaction.go
@@ -1,0 +1,7 @@
+package main
+
+type Interaction struct {
+	ItemName string
+	EntityName string
+	Event *Event
+}
