+++ main.go
@@ -9,16 +9,6 @@ import (
 	"strings"
 )
 
-type Display interface {
-	Show( text string )
- }
-
- type ConsoleDisplay struct {}
-
- func (c ConsoleDisplay) Show(text string) {
-	fmt.Println(text)
- }
-
 func clearScreen() {
     var cmd *exec.Cmd
     if runtime.GOOS == "windows" {
@@ -35,311 +25,16 @@ type Describable interface {
 	GetDescription() string
 }
 
-type Item struct {
-	Name string
-	Description string
-	Weight int
-	Hidden bool
-}
-
-func (i *Item) SetDescription(description string) {
-    i.Description = description
-}
-
-func (i *Item) GetDescription() string {
-    return i.Description
-}
-
-type Room struct {
-	Name string
-	Description string
-	Exits map[string]*Room
-	Items map[string]*Item
-	Entities map[string]*Entity
-}
-
-func (r *Room) SetDescription(description string) {
-    r.Description = description
-}
-
-func (r *Room) GetDescription() string {
-    return r.Description
-}
-
-type Player struct {
-	CurrentRoom *Room
-	Inventory   map[string]*Item
-	CurrentEntity *Entity
-	CarriedWeight int
-	AvailableWeight int
-}
-
-type Entity struct {
-	Name string
-	Description string
-	Hidden bool
-}
-
-type Event struct {
-	Description string
-	Outcome string
-	Triggered bool
-}
-
-type Interaction struct {
-	ItemName string
-	EntityName string
-	Event *Event
-}
-
-var validInteractions = []*Interaction{}
-
-func (e *Entity) SetDescription(description string) {
-    e.Description = description
-}
-
-func (e *Entity) GetDescription() string {
-    return e.Description
-}
-
-func (p *Player) Move(direction string, d Display) {
-	if p.CurrentEntity != nil {
-		p.CurrentEntity = nil
-	}
-	if newRoom, ok := p.CurrentRoom.Exits[direction]; ok {
-		p.CurrentRoom = newRoom
-
-		d.Show(fmt.Sprintf("You are in %s\n", p.CurrentRoom.Name))
-	} else {
-		d.Show("You can't go that way!\n")
-	}
-}
-
-var plateOrder = []string{"first-plate", "second-plate", "third-plate", "fourth-plate", "fifth-plate", "sixth-plate"}
-var currentPlateIndex = 0
-var gameOver = false
-
-
-func (p *Player) Take(itemName string, d Display) {
-	item, ok := p.CurrentRoom.Items[itemName]
-	switch {
-	case !ok || item.Hidden:
-		d.Show(fmt.Sprintf("You can't take %s\n", itemName))
-		return
-	case p.AvailableWeight < item.Weight:
-		d.Show(fmt.Sprintln("Weight limit reached! Please drop an item before taking more."))
-		return
-	case isPlate(itemName):
-		if itemName == plateOrder[currentPlateIndex] {
-			p.Inventory[item.Name] = item
-			p.ChangeCarriedWeight(item, "increase")
-			delete(p.CurrentRoom.Items, item.Name)
-			currentPlateIndex++
-
-			d.Show(fmt.Sprintf("%s has been added to your inventory.\n", item.Name))
-		} else {
-			d.Show(fmt.Sprintln("As you attempt to grab the greasy plates without removing the ones stacked above them, they slip from your grasp and shatter, creating a chaotic mess.\n\nNow Rosie is very grumpy."))
-			gameOver = true
-		}
-		
-	default:
-		p.Inventory[item.Name] = item
-		p.ChangeCarriedWeight(item, "increase")
-		delete(p.CurrentRoom.Items, item.Name)
-
-		d.Show(fmt.Sprintf("%s has been added to your inventory.\n", item.Name))
-	}
-}
-
-func isPlate(itemName string) bool {
-    for _, plate := range plateOrder {
-        if itemName == plate {
-            return true
-        }
-    }
-    return false
-}
-
-func (p *Player) ChangeCarriedWeight(item *Item, operation string) {
-	switch {
-	case operation == "increase":
-		p.CarriedWeight += item.Weight
-		p.AvailableWeight -= item.Weight
-		return
-	case operation == "decrease":
-		p.CarriedWeight -= item.Weight
-		p.AvailableWeight += item.Weight
-		return
-	}
-}
-
-func (p *Player) Drop(itemName string, d Display) {
-	if item, ok := p.Inventory[itemName]; ok {
-		if isPlate(itemName) {
-			d.Show("You can't just leave those plates lying around! It's time to load them into the dishwasher!")
-			return
-		}
-
-		delete(p.Inventory, item.Name)
-		p.ChangeCarriedWeight(item, "decrease")
-		p.CurrentRoom.Items[item.Name] = item
-
-		d.Show(fmt.Sprintf("You dropped %s.\n\n", item.Name))
-	} else {
-		d.Show(fmt.Sprintf("You don't have %s.\n\n", itemName))
-	}
-}
-
-func (p *Player) ShowInventory(d Display) {
-	if len(p.Inventory) == 0 {
-		d.Show(fmt.Sprintf("Your inventory is empty.\nAvailable space: %d\n", p.AvailableWeight))
-		return
-	}
-	d.Show(fmt.Sprintf("Available space: %d\nYour inventory contains:\n", p.AvailableWeight))
-	for itemName, item := range p.Inventory {
-		d.Show(fmt.Sprintf("- %s: %s Weight: %d\n", itemName, item.Description, item.Weight))
-	}
-}
-
-func (p *Player) ShowRoom(d Display) {
-    d.Show(fmt.Sprintf("You are in %s\n\n%s\n", p.CurrentRoom.Name, p.CurrentRoom.Description))
-
-	if p.EntitiesArePresent() {
-			d.Show("\nYou can approach:\n")
-			for _, entity := range p.CurrentRoom.Entities {
-				switch {
-				case p.CurrentEntity != nil:
-					if entity.Name == p.CurrentEntity.Name {
-						d.Show(fmt.Sprintf("- %s (currently approached)\n", entity.Name))
-					} else if !entity.Hidden{
-						d.Show(fmt.Sprintf("- %s\n", entity.Name))
-					}
-				default:
-					if !entity.Hidden{
-						d.Show(fmt.Sprintf("- %s\n", entity.Name))
-					}
-				}
-			}
-		}
-		
-	if p.ItemsArePresent() {
-			d.Show("\nThe room contains:")
-			for itemName, item := range p.CurrentRoom.Items {
-				if !item.Hidden {
-					d.Show(fmt.Sprintf("\n- %s: %s Weight: %d\n", itemName, item.Description, item.Weight))
-				}
-			}
-		}
-	}
-
-	
-
-func (p *Player) ItemsArePresent() bool {
-	if len(p.CurrentRoom.Items) != 0 {
-	for _, item := range p.CurrentRoom.Items {
-		if !item.Hidden {
-			return true
-		}
-	}
-}
-	return false
-}
-
-func (p *Player) EntitiesArePresent() bool {
-	if len(p.CurrentRoom.Entities) != 0 {
-	for _, entity := range p.CurrentRoom.Entities {
-		if !entity.Hidden {
-			return true
-		}
-	}
-}
-	return false
-}
-
-
-func (p *Player) Approach(entityName string, d Display) {
-	if p.CurrentEntity != nil {
-		p.CurrentEntity = nil
-	}
-	if entity, ok := p.CurrentRoom.Entities[entityName]; ok && !entity.Hidden{
-
-		p.CurrentEntity = entity
-		d.Show(entity.Description)
-	} else {
-		d.Show(fmt.Sprintf("You can't approach %s.\n", entityName))
-	}
-}
-
-func (p *Player) Leave() {
-	if p.CurrentEntity != nil {
-		p.CurrentEntity = nil
-		p.ShowRoom(ConsoleDisplay{})
-	} else {
-		fmt.Println("You have not approached anything. If you wish to leave the game, use the exit command.")
-	}
-}
-
-func (p *Player) ShowMap(d Display) {
-	for direction, exit := range p.CurrentRoom.Exits {
-		d.Show(fmt.Sprintf("%s: %s\n", direction, exit.Name))
-	}
-}
-
-func (p *Player) Use(itemName string, target string, d Display) {
-
-	if p.CurrentEntity == nil {
-		d.Show("Approach to use an item.\n")
-		return
-	}
-
-	if p.CurrentEntity.Name != target {
-		d.Show(fmt.Sprintf("%s not found.\n", target))
-		return
-	}
-
-	if itemIsNotInInventory(p, itemName) {
-		d.Show(fmt.Sprintf("You don't have %s.\n", itemName))
-		return
-	}
-	
-	for _, interaction := range validInteractions {
-		if interactionIsValid(interaction, itemName, target) {
-			handleInteraction(p, interaction, itemName)
-			return
-		} 
-	}
-	d.Show(fmt.Sprintf("You can't use %s on %s.\n", itemName, target))
-}
-
-
-func itemIsNotInInventory(p *Player, itemName string) bool {
-	_, ok := p.Inventory[itemName]
-    return !ok
-}
-func interactionIsValid(interaction *Interaction, itemName string, target string) bool {
-	return interaction.ItemName == itemName && interaction.EntityName == target
-}
-
-func handleInteraction(p *Player, interaction *Interaction, itemName string) {
-	p.TriggerEvent(interaction.Event)
-	p.ChangeCarriedWeight(p.Inventory[itemName], "decrease")
-	delete(p.Inventory, itemName)
-}
-
-func (p *Player) TriggerEvent(event *Event) {
-	fmt.Println(event.Outcome)
-	event.Triggered = true
-}
-
-func updateDescription(d Describable, newDescription string) {
-	d.SetDescription(newDescription)
-}
-
 func showCommands(d Display) {
 	d.Show("-exit -> quits the game\n\n-commands -> shows the commands\n\n-look -> shows the content of the room.\n\n-approach <entity> -> to approach an entity\n\n-leave -> to leave an entity\n\n-inventory -> shows items in the inventory\n\n-take <item> -> to take an item into your inventory\n\n-drop <item> -> to drop an item from your inventory and move it to the current room\n\n-use <item> -> to make use of a certain item when you approach an entity\n\n-move <direction> -> to move to a different room\n\n-map -> shows the directions you can take\n")
 }
 
+var validInteractions = []*Interaction{}
+
+var gameOver = false
+
 func main() {
+
 	introduction := "It's the last day at the Academy, and you and your fellow graduates are ready to take on the final hack-day challenge.\nHowever, this time, it's different. Alan and Dan, your instructors, have prepared something more intense than ever before — a true test of your problem-solving and coding skills.\nThe doors to the academy are locked, the windows sealed. The only way out is to find and solve a series of riddles that lead to the terminal in a hidden room.\nThe challenge? Crack the code on the terminal to unlock the doors. But it's not that simple.\nYou'll need to gather items, approach Alan and Dan for cryptic tips, and outsmart the obstacles they've laid out for you.\nAs the tension rises, only your wits, teamwork, and knowledge can guide you to freedom.\nAre you ready to escape?\nOh and remember... You don't want to make Rosie grumpy! So don't do anything crazy.\n\nif at any point you feel lost, type 'commands' to display the list of all commands.\nThe command 'look' is always useful to get your bearings and see the options available to you.\nThe command 'exit' will make you quit the game at any time. Make sure you do mean to use it, or you will inadvertently lose all of your progress!"
 
 	introductionShown:= false
